documentation update; better reporting on dbbackup and dbutil
