+++ docs/Development/DeveloperSetup.rst
@@ -65,10 +65,16 @@ Local storage mode is the default. Leave ``AWS_REGION`` unset, or set it to
 .. code-block:: bash
 
    make show-storage-mode
+   eval "$(make show-local-vars)"
    make start-local-services
    make make-local-demo
    make run-local-lambda-debug
    make run-local-debug
+
+``show-local-vars`` prints shell commands for the complete non-demo local debug
+environment. Evaluating its output configures subsequent direct commands, such
+as ``poetry run dbutil report``, to use DynamoDB Local and MinIO. It also clears
+AWS profiles and demo-mode variables so they cannot override the local setup.
 
 Remote storage mode runs local Flask and, if needed, the local lambda debug
 bridge against real AWS S3 and DynamoDB. Use this only with a development
@@ -139,7 +145,7 @@ Create A Course
 
 .. code-block:: bash
 
-   AWS_REGION=local poetry run python src/dbutil.py create-course \
+   AWS_REGION=local poetry run dbutil create-course \
      --course_name "Test Course" \
      --course_id "test" \
      --admin_email admin@example.com \
