+++ docs/Development/DynamoDB.rst
@@ -23,9 +23,15 @@ recommend using the version downloaded as a JAR file.
 The canonical table model is split between the application item schemas and the table creation
 contract. Item fields are defined in ``src/app/schema.py``, and attribute-name constants are
 defined at the top of ``src/app/odb.py``. DynamoDB table definitions are in
-``etc/dynamodb_tables.json``. ``src/dbutil.py createdb`` creates the tables from that JSON file
+``etc/dynamodb_tables.json``. ``poetry run dbutil createdb`` creates the tables from that JSON file
 and populates the demo course; ``odbmaint.drop_tables()`` drops the configured tables for local
 test and reset workflows.
+
+Use ``poetry run dbutil list-prefixes`` to list complete Plant Tracer table
+prefixes available through the current DynamoDB connection settings. Other
+``dbutil`` commands require ``DYNAMODB_TABLE_PREFIX``; when it is missing, they
+print the active ``AWS_REGION`` and the same available-prefix list before
+exiting.
 
 The SAM/CloudFormation stack does not own DynamoDB tables. For an existing stack that still owns
 tables from an older template, retain or otherwise detach those table resources before deploying a
