+++ src/dbbackup.py
@@ -34,6 +34,13 @@ from app import mailer
 from app import odb
 from app import odbmaint
 from app.constants import C, __version__
+from app.dynamodb_prefixes import (
+    DYNAMODB_TABLE_CONFIGURATIONS,
+    PrefixSummary,
+    REQUIRED_PREFIX_TABLES,
+    format_list_prefixes,
+    prefix_summaries,
+)
 from app.odb import (
     ADMIN_FOR_COURSES,
     ADMINS_FOR_COURSE,
@@ -94,15 +101,8 @@ MOVIE_S3_URN_FIELDS = (MOVIE_DATA_URN, MOVIE_ZIPFILE_URN, MOVIE_TRACED_URN)
 USER_DATE_FIELDS = (CREATED,)
 COURSE_DATE_FIELDS = (CREATED, CREATED_AT)
 MOVIE_DATE_FIELDS = (CREATED_AT, DATE_UPLOADED, MOVIE_STATE_UPDATED_AT)
-LIST_PREFIX_HEADERS = ("prefix", "courses", "users", "movies", "from", "to")
-LIST_PREFIX_RIGHT_ALIGNED = (False, True, True, True, False, False)
 TRACKPOINTS = "trackpoints"
 TRACKPOINT_LABEL = "label"
-DYNAMODB_TABLE_CONFIGURATIONS = tuple(odbmaint.load_table_configurations())
-REQUIRED_PREFIX_TABLES = tuple(
-    table_config[odbmaint.TableName]
-    for table_config in DYNAMODB_TABLE_CONFIGURATIONS
-)
 
 
 class DbBackupError(RuntimeError):
@@ -265,44 +265,6 @@ class BackupDataset(BaseModel):
         }
 
 
-class PrefixSummary(BaseModel):
-    """One complete DynamoDB table-prefix summary."""
-
-    prefix: str
-    courses: int
-    users: int
-    movies: int
-    date_from: int | None = None
-    date_to: int | None = None
-
-
-class TimestampRange(BaseModel):
-    """A compact inclusive range of Unix epoch seconds."""
-
-    date_from: int | None = None
-    date_to: int | None = None
-
-    def include(self, value: Any) -> None:
-        epoch = epoch_seconds(value)
-        if epoch is None:
-            return
-        if self.date_from is None or epoch < self.date_from:
-            self.date_from = epoch
-        if self.date_to is None or epoch > self.date_to:
-            self.date_to = epoch
-
-    def include_range(self, other: "TimestampRange") -> None:
-        self.include(other.date_from)
-        self.include(other.date_to)
-
-
-class TableDateSummary(BaseModel):
-    """Count and timestamp range for a table scan."""
-
-    count: int = 0
-    date_range: TimestampRange
-
-
 def encode_attribute_value(attribute_value: dict[str, Any]) -> dict[str, Any]:
     """Make a DynamoDB AttributeValue JSON-serializable."""
     if "B" in attribute_value:
@@ -413,45 +375,6 @@ def query_all(table, **query_kwargs) -> list[dict[str, Any]]:
         query_kwargs["ExclusiveStartKey"] = last_key
 
 
-def list_dynamodb_table_names(dynamodb) -> list[str]:
-    table_names: list[str] = []
-    list_kwargs: dict[str, Any] = {}
-    while True:
-        response = dynamodb.meta.client.list_tables(**list_kwargs)
-        table_names.extend(response.get("TableNames", []))
-        last_table_name = response.get("LastEvaluatedTableName")
-        if last_table_name is None:
-            return sorted(table_names)
-        list_kwargs["ExclusiveStartTableName"] = last_table_name
-
-
-def prefix_for_table_name(table_name: str, base_table_name: str) -> str | None:
-    if table_name == base_table_name:
-        return ""
-    if not table_name.endswith(base_table_name):
-        return None
-    prefix = table_name[: -len(base_table_name)]
-    if prefix.endswith("-"):
-        return prefix
-    return None
-
-
-def complete_prefixes(table_names: list[str]) -> list[str]:
-    tables_by_prefix: dict[str, set[str]] = {}
-    for table_name in table_names:
-        for base_table_name in REQUIRED_PREFIX_TABLES:
-            prefix = prefix_for_table_name(table_name, base_table_name)
-            if prefix is not None:
-                tables_by_prefix.setdefault(prefix, set()).add(base_table_name)
-
-    required_tables = set(REQUIRED_PREFIX_TABLES)
-    return sorted(
-        prefix
-        for prefix, base_table_names in tables_by_prefix.items()
-        if base_table_names >= required_tables
-    )
-
-
 def required_table_names_for_prefix(table_prefix: str) -> set[str]:
     normalized_prefix = normalized_table_prefix(table_prefix)
     return {
@@ -483,127 +406,6 @@ def restore_target_state(dynamodb, table_prefix: str) -> RestoreTargetState:
         existing_table_names=existing_table_names,
         missing_table_names=missing_table_names,
     )
-
-
-def epoch_seconds(value: Any) -> int | None:
-    """Return a sane Unix timestamp, or None for absent/non-timestamp values."""
-    if value is None or isinstance(value, bool):
-        return None
-    if isinstance(value, Decimal):
-        try:
-            epoch = int(value)
-        except (OverflowError, ValueError):
-            return None
-    elif isinstance(value, int):
-        epoch = value
-    elif isinstance(value, str):
-        try:
-            epoch = int(Decimal(value.strip()))
-        except (ArithmeticError, ValueError):
-            return None
-    else:
-        return None
-    if epoch <= 0:
-        return None
-    return epoch
-
-
-def date_projection_kwargs(date_fields: tuple[str, ...]) -> dict[str, Any]:
-    expression_names = {
-        f"#date{i}": field_name
-        for i, field_name in enumerate(date_fields)
-    }
-    return {
-        "ProjectionExpression": ", ".join(expression_names),
-        "ExpressionAttributeNames": expression_names,
-    }
-
-
-def summarize_table_dates(table, date_fields: tuple[str, ...]) -> TableDateSummary:
-    summary = TableDateSummary(date_range=TimestampRange())
-    scan_kwargs: dict[str, Any] = {
-        "ConsistentRead": True,
-        **date_projection_kwargs(date_fields),
-    }
-    while True:
-        response = table.scan(**scan_kwargs)
-        summary.count += int(response["Count"])
-        for item in response.get("Items", []):
-            for field_name in date_fields:
-                summary.date_range.include(item.get(field_name))
-        last_key = response.get("LastEvaluatedKey")
-        if last_key is None:
-            return summary
-        scan_kwargs["ExclusiveStartKey"] = last_key
-
-
-def format_epoch_seconds(epoch: int | None) -> str:
-    if epoch is None:
-        return "-"
-    return time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime(epoch))
-
-
-def list_prefix_row(summary: PrefixSummary) -> tuple[str, str, str, str, str, str]:
-    prefix = summary.prefix if summary.prefix else "(none)"
-    return (
-        prefix,
-        str(summary.courses),
-        str(summary.users),
-        str(summary.movies),
-        format_epoch_seconds(summary.date_from),
-        format_epoch_seconds(summary.date_to),
-    )
-
-
-def format_list_prefix_row(
-    row: tuple[str, ...],
-    widths: tuple[int, ...],
-) -> str:
-    cells = []
-    for cell, width, right_aligned in zip(row, widths, LIST_PREFIX_RIGHT_ALIGNED):
-        cells.append(cell.rjust(width) if right_aligned else cell.ljust(width))
-    return "  ".join(cells).rstrip()
-
-
-def format_list_prefixes(summaries: list[PrefixSummary]) -> list[str]:
-    rows = [LIST_PREFIX_HEADERS] + [list_prefix_row(summary) for summary in summaries]
-    widths = tuple(
-        max(len(row[index]) for row in rows)
-        for index in range(len(LIST_PREFIX_HEADERS))
-    )
-    return [format_list_prefix_row(row, widths) for row in rows]
-
-
-def prefix_summaries(dynamodb) -> list[PrefixSummary]:
-    summaries: list[PrefixSummary] = []
-    for prefix in complete_prefixes(list_dynamodb_table_names(dynamodb)):
-        course_summary = summarize_table_dates(
-            dynamodb.Table(prefix + COURSES),
-            COURSE_DATE_FIELDS,
-        )
-        user_summary = summarize_table_dates(
-            dynamodb.Table(prefix + USERS),
-            USER_DATE_FIELDS,
-        )
-        movie_summary = summarize_table_dates(
-            dynamodb.Table(prefix + MOVIES),
-            MOVIE_DATE_FIELDS,
-        )
-        date_range = TimestampRange()
-        date_range.include_range(course_summary.date_range)
-        date_range.include_range(user_summary.date_range)
-        date_range.include_range(movie_summary.date_range)
-        summaries.append(
-            PrefixSummary(
-                prefix=prefix,
-                courses=course_summary.count,
-                users=user_summary.count,
-                movies=movie_summary.count,
-                date_from=date_range.date_from,
-                date_to=date_range.date_to,
-            )
-        )
-    return summaries
 
 
 def prompt_run_aws_sso_login(exc: TokenRetrievalError) -> bool:
